Remove redundant casts to int from wxGrid::Redimension() code
There is no need to cast already int variables to int before negating
them (this was probably a left over from 25 year old initial version
where these variables might have been unsigned).

+++ src/generic/grid.cpp
@@ -3476,11 +3476,11 @@ bool wxGrid::Redimension( wxGridTableMessage& msg )
             UpdateCurrentCellOnRedim();
 
             if ( m_selection )
-                m_selection->UpdateRows( pos, -((int)numRows) );
+                m_selection->UpdateRows( pos, -numRows );
             wxGridCellAttrProvider * attrProvider = m_table->GetAttrProvider();
             if (attrProvider)
             {
-                attrProvider->UpdateAttrRows( pos, -((int)numRows) );
+                attrProvider->UpdateAttrRows( pos, -numRows );
 
 // ifdef'd out following patch from Paul Gammans
 #if 0
@@ -3665,11 +3665,11 @@ bool wxGrid::Redimension( wxGridTableMessage& msg )
             UpdateCurrentCellOnRedim();
 
             if ( m_selection )
-                m_selection->UpdateCols( pos, -((int)numCols) );
+                m_selection->UpdateCols( pos, -numCols );
             wxGridCellAttrProvider * attrProvider = m_table->GetAttrProvider();
             if (attrProvider)
             {
-                attrProvider->UpdateAttrCols( pos, -((int)numCols) );
+                attrProvider->UpdateAttrCols( pos, -numCols );
 
 // ifdef'd out following patch from Paul Gammans
 #if 0
